+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.7.11] - 2016-06-10
 ---------------------
 ##### Fixed
 - In IE11, external data cannot be added to model.
+- If instanceID node is missing, no error is output to user.
 
 [4.7.10] - 2016-06-08
 ---------------------